+++ src/context/AdminProviders/UsersContext.jsx
@@ -1,6 +1,11 @@
 import { createContext, useContext } from 'react'
-import useEntityManagement from '../hooks/useEntityManagement'
-import { getUsers, addUser, updateUser, deleteUser } from '../services/Usuarios'
+import useEntityManagement from '../../hooks/useEntityManagement'
+import {
+	getUsers,
+	addUser,
+	updateUser,
+	deleteUser,
+} from '../../services/Usuarios'
 
 export const UsersContext = createContext()
