+++ src/pages/Admin/pages/Routes/Routes.jsx
@@ -1,7 +1,7 @@
 import PageSchema from '../PageSchema/PageSchema'
 import ModalAdmin from '../../components/ModalAdmin/ModalAdmin'
 import Error from '../../../../components/Error/Error'
-import { useRoutes } from '../../../../context/RoutesContext'
+import { useRoutes } from '../../../../context/AdminProviders'
 import { useState } from 'react'
 
 function RoutesAdmin() {
